+++ documentation/stories/policy/form.jsx
@@ -7,6 +7,7 @@ import { createTheme, ThemeProvider } from '@mui/material/styles';
 import { orange } from '@mui/material/colors';
 import "../../../src/i18n";
 import React from 'react';
+import { SnackbarProvider } from "notistack";
 
 /**
  * Primary UI component for user interaction
@@ -31,6 +32,7 @@ export const PolicyMainForm = ({
         return "Tenant1"
     }
     return (
+        <SnackbarProvider maxSnack={5}>
         <ThemeProvider theme={theme}>
             <BrowserRouter>
                 <Grid
@@ -48,6 +50,7 @@ export const PolicyMainForm = ({
                 </Grid>
             </BrowserRouter>
         </ThemeProvider>
+        </SnackbarProvider>
     );
 };