+++ documentation/stories/service/form.jsx
@@ -7,6 +7,7 @@ import { createTheme, ThemeProvider} from '@mui/material/styles';
 import { orange } from '@mui/material/colors';
 import "../../../src/i18n";
 import React from 'react';
+import { SnackbarProvider } from "notistack";
 
 /**
  * Primary UI component for user interaction
@@ -26,6 +27,7 @@ export const ServiceMainForm = ({
     tenantName_id,
 }) => {
     return (
+        <SnackbarProvider maxSnack={5}>
         <ThemeProvider theme={theme}>
             <BrowserRouter>
                 <Grid
@@ -41,6 +43,7 @@ export const ServiceMainForm = ({
                 </Grid>
             </BrowserRouter>
         </ThemeProvider>
+        </SnackbarProvider>
     );
 };